Search and EntrySummary refactor
Plus other TODO tidying up
